rename IndexService to QueryService
