--- packages/core/protocols/src/proto/dxos/agent/query.proto
@@ -1,28 +1,0 @@
-//
-// Copyright 2023 DXOS.org
-//
-
-syntax = "proto3";
-
-import "dxos/echo/filter.proto";
-import "dxos/keys.proto";
-import "dxos/echo/object.proto";
-
-package dxos.agent.query;
-
-message QueryRequest {
-  optional string queryId = 1;
-  filter.Filter filter = 2;
-}
-
-message QueryResult {
-  string id = 1;
-  keys.PublicKey spaceKey = 2;
-  float rank = 3;
-}
-
-message QueryResponse {
-  optional string queryId = 1;
-  repeated QueryResult results = 2;
-  repeated object.EchoObject objects = 3;
-}
