+++ packages/sdk/client/src/echo/space-list.ts
@@ -172,7 +172,7 @@ export class SpaceList extends MulticastObservable<Space[]> implements Echo {
       this._ctx.onDispose(() => agentQuerySourceProvider.close());
 
       this._graph.registerQuerySourceProvider(
-        new IndexQuerySourceProvider({ echo: this, service: this._serviceProvider.services.IndexService! }),
+        new IndexQuerySourceProvider({ echo: this, service: this._serviceProvider.services.QueryService! }),
       );
       subscription.unsubscribe();
     });
@@ -186,7 +186,7 @@ export class SpaceList extends MulticastObservable<Space[]> implements Echo {
   }
 
   async setIndexConfig(config: IndexConfig) {
-    await this._serviceProvider.services.IndexService?.setConfig(config, { timeout: 20_000 }); // TODO(dmaretskyi): Set global timeout instead.
+    await this._serviceProvider.services.QueryService?.setIndexConfig(config, { timeout: 20_000 }); // TODO(dmaretskyi): Set global timeout instead.
   }
 
   /**
